changes in attackfactory: * attackfactory creates abstract IAttacks rather than concrete attacks. * attackfactory ensures that the owner is set directly, avoiding a null exception on the owner.
added new attacks

created a new concrete of IAttack : Defend, replacing the defend() method. Simply put, it's now a real attack and not a different behavior :
Pros : more simple to handle.
Cons : "defend" is now an "attack", which is semantically a bit wrong.

create AI structure and first behaviors :
- Random AI
- score-based AI : calculate a score based on damage status and kill potential of each attack.

fixed a bug that reversed the stat variation in the animal factory